Separating Rewards NTP redux code
