+++ components/brave_new_tab_ui/actions/new_tab_actions.ts
@@ -10,7 +10,7 @@ import { types } from '../constants/new_tab_types'
 import { Preferences } from '../api/preferences'
 import { Stats } from '../api/stats'
 import { PrivateTabData } from '../api/privateTabData'
-import { InitialData, InitialRewardsData, PreInitialRewardsData } from '../api/initialData'
+import { InitialData } from '../api/initialData'
 
 export const statsUpdated = (stats: Stats) =>
   action(types.NEW_TAB_STATS_UPDATED, {
@@ -31,65 +31,10 @@ export const preferencesUpdated = (preferences: Preferences) =>
 export const setInitialData = (initialData: InitialData) =>
   action(types.NEW_TAB_SET_INITIAL_DATA, initialData)
 
-export const createWallet = () => action(types.CREATE_WALLET, {})
-
-export const onEnabledMain = (enabledMain: boolean, enabledAds?: boolean) => action(types.ON_ENABLED_MAIN, {
-  enabledMain,
-  enabledAds
-})
-
-export const onAdsEnabled = (enabled: boolean) => action(types.ON_ADS_ENABLED, {
-  enabled
-})
-
-export const onWalletInitialized = (result: NewTab.RewardsResult) => action(types.ON_WALLET_INITIALIZED, {
-  result
-})
-
-export const onAdsEstimatedEarnings = (amount: number) => action(types.ON_ADS_ESTIMATED_EARNINGS, {
-  amount
-})
-
-export const onBalanceReport = (properties: {month: number, year: number, report: NewTab.RewardsBalanceReport}) => action(types.ON_BALANCE_REPORT, {
-  month: properties.month,
-  year: properties.year,
-  report: properties.report
-})
-
-export const onPromotions = (result: number, promotions: NewTab.Promotion[]) => action(types.ON_PROMOTIONS, {
-  result,
-  promotions
-})
-
-export const dismissNotification = (id: string) => action(types.DISMISS_NOTIFICATION, {
-  id
-})
-
-export const onBalance = (balance: NewTab.RewardsBalance) => action(types.ON_BALANCE, {
-  balance
-})
-
-export const onWalletExists = (exists: boolean) => action(types.ON_WALLET_EXISTS, {
-  exists
-})
-
-export const setInitialRewardsData = (initialRewardsData: InitialRewardsData) => action(types.SET_INITIAL_REWARDS_DATA, initialRewardsData)
-
-export const setPreInitialRewardsData = (preInitialRewardsData: PreInitialRewardsData) => action(types.SET_PRE_INITIAL_REWARDS_DATA, preInitialRewardsData)
-
-export const onPromotionFinish = (result: NewTab.RewardsResult, promotion: NewTab.Promotion) => action(types.ON_PROMOTION_FINISH, {
-  result,
-  promotion
-})
-
 export const removeStackWidget = (widget: NewTab.StackWidget) => action(types.REMOVE_STACK_WIDGET, {
   widget
 })
 
 export const setForegroundStackWidget = (widget: NewTab.StackWidget) => action(types.SET_FOREGROUND_STACK_WIDGET, {
   widget
-})
-
-export const setOnlyAnonWallet = (onlyAnonWallet: boolean) => action(types.SET_ONLY_ANON_WALLET, {
-  onlyAnonWallet
 })
