+++ components/brave_new_tab_ui/api/getActions.ts
@@ -7,17 +7,18 @@ import { bindActionCreators } from 'redux'
 import * as newTabActions from '../actions/new_tab_actions'
 import * as gridSitesActions from '../actions/grid_sites_actions'
 import * as binanceActions from '../actions/binance_actions'
+import * as rewardsActions from '../actions/rewards_actions'
 import store from '../store'
 
 /**
  * Get actions from the C++ back-end down to front-end components
  */
-let actions: typeof newTabActions & typeof gridSitesActions & typeof binanceActions
+let actions: typeof newTabActions & typeof gridSitesActions & typeof binanceActions & typeof rewardsActions
 export default function getActions () {
   if (actions) {
     return actions
   }
-  const allActions = Object.assign({}, newTabActions, gridSitesActions, binanceActions)
+  const allActions = Object.assign({}, newTabActions, gridSitesActions, binanceActions, rewardsActions)
   actions = bindActionCreators(allActions, store.dispatch.bind(store))
   return actions
 }
