+++ components/brave_new_tab_ui/constants/new_tab_types.ts
@@ -10,23 +10,7 @@ export const enum types {
   NEW_TAB_DISMISS_BRANDED_WALLPAPER_NOTIFICATION = '@@newtab/NEW_TAB_DISMISS_BRANDED_WALLPAPER_NOTIFICATION',
   NEW_TAB_SET_INITIAL_DATA = '@@newtab/NEW_TAB_SET_INITIAL_DATA',
   REMOVE_STACK_WIDGET = '@@newtab/REMOVE_STACK_WIDGET',
-  SET_FOREGROUND_STACK_WIDGET = '@@newtab/SET_FOREGROUND_STACK_WIDGET',
-  // Rewards Widget
-  CREATE_WALLET = '@@newtab/CREATE_WALLET',
-  DISMISS_NOTIFICATION = '@@newtab/DISMISS_NOTIFICATION',
-  ON_ADS_ENABLED = '@@newtab/ON_ADS_ENABLED',
-  ON_ADS_ESTIMATED_EARNINGS = '@@newtab/ON_ADS_ESTIMATED_EARNINGS',
-  ON_ENABLED_MAIN = '@@newtab/ON_ENABLED_MAIN',
-  ON_WALLET_INITIALIZED = '@@newtab/ON_WALLET_INITIALIZED',
-  ON_BALANCE_REPORT = '@@newtab/ON_BALANCE_REPORT',
-  ON_PROMOTIONS = '@@newtab/ON_PROMOTIONS',
-  ON_PROMOTION_FINISH = '@@newtab/ON_PROMOTION_FINISH',
-  ON_BALANCE = '@@newtab/ON_BALANCE',
-  ON_WALLET_EXISTS = '@@newtab/ON_WALLET_EXISTS',
-  SET_INITIAL_REWARDS_DATA = '@@newtab/SET_INITIAL_REWARDS_DATA',
-  SET_PRE_INITIAL_REWARDS_DATA = '@@newtab/SET_PRE_INITIAL_REWARDS_DATA',
-  ON_WIDGET_POSITION_CHANGED = '@@newtab/ON_WIDGET_POSITION_CHANGED',
-  SET_ONLY_ANON_WALLET = '@@newtab/SET_ONLY_ANON_WALLET'
+  SET_FOREGROUND_STACK_WIDGET = '@@newtab/SET_FOREGROUND_STACK_WIDGET'
 }
 
 export type DismissBrandedWallpaperNotificationPayload = {
