+++ components/brave_new_tab_ui/containers/app.tsx
@@ -15,10 +15,11 @@ import NewTabPage from './newTab'
 import * as newTabActions from '../actions/new_tab_actions'
 import * as gridSitesActions from '../actions/grid_sites_actions'
 import * as binanceActions from '../actions/binance_actions'
+import * as rewardsActions from '../actions/rewards_actions'
 import * as PreferencesAPI from '../api/preferences'
 
 interface Props {
-  actions: typeof newTabActions & typeof gridSitesActions & typeof binanceActions
+  actions: typeof newTabActions & typeof gridSitesActions & typeof binanceActions & typeof rewardsActions
   newTabData: NewTab.State
   gridSitesData: NewTab.GridSitesState
 }
@@ -57,7 +58,7 @@ const mapStateToProps = (state: NewTab.ApplicationState) => ({
 })
 
 const mapDispatchToProps = (dispatch: Dispatch) => {
-  const allActions = Object.assign({}, newTabActions, gridSitesActions, binanceActions)
+  const allActions = Object.assign({}, newTabActions, gridSitesActions, binanceActions, rewardsActions)
   return {
     actions: bindActionCreators(allActions, dispatch)
   }
