+++ components/brave_new_tab_ui/store.ts
@@ -10,6 +10,6 @@ import { createStore, applyMiddleware } from 'redux'
 import thunk from 'redux-thunk'
 
 // Feature core reducer
-import rootReducer from './reducers'
+import { mainNewTabReducer } from './reducers'
 
-export default createStore(rootReducer, applyMiddleware(thunk))
+export default createStore(mainNewTabReducer, applyMiddleware(thunk))
