+++ components/test/brave_new_tab_ui/api/data_test.ts
@@ -6,13 +6,15 @@ import getActions from '../../../brave_new_tab_ui/api/getActions'
 import { getTopSites } from '../../../brave_new_tab_ui/api/topSites'
 import * as newTabActions from '../../../brave_new_tab_ui/actions/new_tab_actions'
 import * as topSitesActions from '../../../brave_new_tab_ui/actions/grid_sites_actions'
+import * as binanceActions from '../../../brave_new_tab_ui/actions/binance_actions'
+import * as rewardsActions from '../../../brave_new_tab_ui/actions/rewards_actions'
 import { types as topSitesTypes } from '../../../brave_new_tab_ui/constants/grid_sites_types'
 
 describe('new tab data api tests', () => {
   describe('getActions', () => {
     it('returns an object with the same keys mimicking the original new tab actions', () => {
       const assertion = getActions()
-      const actions = Object.assign({}, newTabActions, topSitesActions)
+      const actions = Object.assign({}, newTabActions, topSitesActions, binanceActions, rewardsActions)
       expect(Object.keys(assertion)).toEqual(Object.keys(actions))
     })
     it('can call an action from getActions', () => {
